requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
loguru==0.7.2
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pydantic==1.10.12
pyotp==2.9.0
pytest==8.2.2
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.2
typing_extensions==4.14.0
urllib3==2.4.0
websocket-client==1.8.0
win32_setctime==1.2.0
